requirements.txt
flask
openai
python-telegram-bot==13.15
requests